add mvp test to local exec
